requirements.txt
streamlit
ultralytics
pillow
numpy
opencv-python
plotly
pandas
scipy
